requirements.txt
Django==4.2.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
pytz==2023.3.post1
